+++ src/charm.py
@@ -28,7 +28,7 @@ from ops.model import (
     WaitingStatus,
 )
 
-from resource_manager.decorator import InsufficientPermissionError
+from resource_manager.permission import InsufficientPermissionError
 from resource_manager.gateway import CreateGatewayError, GatewayResourceManager
 from resource_manager.resource_manager import InvalidResourceError
 from state.config import CharmConfig, InvalidCharmConfigError
